+++ src/sql/destroys.sql
@@ -1,11 +1,11 @@
 DROP TABLE IF EXISTS destroys;
 CREATE TABLE destroys (
-    action_index INTEGER UNSIGNED NOT NULL, -- Unique action index
-    tick_id      INTEGER UNSIGNED,          -- id of record in index_ticks table
+    action_index BIGINT UNSIGNED NOT NULL, -- Unique action index
+    tick_id      BIGINT UNSIGNED,          -- id of record in index_ticks table
     amount       VARCHAR(250),              -- Amount of token to destroy
-    source_id    INTEGER UNSIGNED,          -- id of record in index_addresses table
-    memo_id      INTEGER UNSIGNED,          -- id of record in index_memos table 
-    status_id    INTEGER UNSIGNED           -- id of record in index_statuses table
+    source_id    BIGINT UNSIGNED,          -- id of record in index_addresses table
+    memo_id      BIGINT UNSIGNED,          -- id of record in index_memos table 
+    status_id    BIGINT UNSIGNED           -- id of record in index_statuses table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index   ON destroys (action_index);
